% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_to_gene.R
\name{add_gex_and_LSI}
\alias{add_gex_and_LSI}
\title{Add gene expression and joint LSI to archr project}
\usage{
add_gex_and_LSI(archr.proj)
}
\arguments{
\item{archr.proj}{}
}
\description{
Add gene expression and joint LSI to archr project
}
\keyword{internal}
