% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Post-pipeline annotation of marker peaks}
\usage{
annotatePeaks(
  marker.table,
  archr.proj,
  dataset,
  groupBy,
  publish,
  filename,
  shiny.dir = NULL,
  ucsc.user = NULL,
  ucsc.session = NULL
)
}
\arguments{
\item{marker.table}{file.path to a tab-seperated file containing results of ArchR marker peak analysis}

\item{archr.proj}{An already setup ArchR project or a string to build a new ArchR project.}

\item{dataset}{Information about the dataset that these marker peaks belong to. E.g. Brain region, etc.}

\item{groupBy}{The annotation level used to group data for peak calling. E.g. Class, Subclass, etc.}

\item{publish}{Set to 'TRUE' if you are ready to make the marker peak table viewable on Shiny.}

\item{filename}{filename to save annotated marker table}

\item{ucsc.user}{A ucsc genome browser user/account which to contains the session to build links from.}

\item{ucsc.session}{A registred session for the UCSC genome browser user/account specified in \code{ucsc.user}.}
}
\value{
annotated marker table
}
\description{
Defines additional marker peak annotations and standardizes reporting tables.
}
