% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Builds the primary archr project object}
\usage{
build(
  arrow.file.dir,
  cell.annotation.file,
  archr.proj = "peakCallingPipeline",
  ...
)
}
\arguments{
\item{arrow.file.dir}{Directory where all the Arrow files are stored}

\item{cell.annotation.file}{File that contains sample_id and additional cell annotations to append to the ArchRProject}

\item{archr.proj}{An already setup ArchR project or a string to build a new ArchR project.}

\item{...}{Additional parameters to ArchRProject()}
}
\value{
None
}
\description{
Sets global variables about specific genome build and number of threads to utilze on the machine.
}
\keyword{internal}
