% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gini_index.R
\name{calculate_gini_index}
\alias{calculate_gini_index}
\title{Function to calculate gini index - returns marker peaks file with gini_index as a new column}
\usage{
calculate_gini_index(
  archr.proj,
  groupBy,
  filename = file.path(getOutputDirectory(archr.proj), "MarkerPeaks", groupBy,
    paste0(groupBy, "_annotated_markerPeaks.tsv")),
  numWorkers = 10,
  max_sample_size = 500
)
}
\arguments{
\item{archr.proj}{ArchR object with peaks called}

\item{groupBy}{Metadata field to group cells by}

\item{filename}{output file for markerpeaks}

\item{max_sample_size}{}
}
\value{
marker peak table
}
\description{
Function to calculate gini index - returns marker peaks file with gini_index as a new column
}
\author{
Marcus Hooper
}
