% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDuplicates.R
\name{checkDuplicates}
\alias{checkDuplicates}
\title{Identify duplicates in a column and tag with each matched element's row number.}
\usage{
checkDuplicates(dataSet, columnName, ...)
}
\arguments{
\item{dataSet}{A tibble or data frame object.}

\item{columnName}{A character vector containing the name of the column to search for
duplicates. NOTE: the column names can be passed with or without quotes, so long as
there are not special characters in the column name.}

\item{\dots}{additional parameters to pass to function. They include:
\itemize{
 \item \code{splitDelim}: A character string that will separate supplied elements into parts to
 identify potential matches within each each delimiter separated string. Default is NULL and the
 string will not be split.
}}
}
\value{
A character vector. If the same value is identified more than once in the same column,
the corresponding row number/index for each matched instance is returned.
}
\description{
Identify duplicates in a column and tag with each matched element's row number.
}
\examples{
# Simple searching for duplicate letters
letterData <- data.frame(letter = sample(letters[1:15], size = 10, replace = TRUE))
letterData$duplicates <- checkDuplicates(dataSet = letterData, columnName = "letter")
letterData

# An example where one row has multiple elements to check duplicates for
colorData <- data.frame(colors = c("PurPle,Green", "blue", "Green", "blue",
"RED","purple"))
colorData$duplicates <- checkDuplicates(dataSet = colorData,
columnName = "colors", splitDelim = ",")
colorData

}
\author{
Ama Nyame-Mensah
}
