% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorData}
\alias{colorData}
\title{Color Data Set}
\format{
A data frame with 6 rows and 1 variable(s):
\describe{
  \item{colors}{basic color names}
}
}
\usage{
colorData
}
\description{
A data set containing a column of colors.
}
\details{
#' Note: this data set is included for illustrative purposes only.

The variables are as follows:
}
\keyword{datasets}
