% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_cols.R
\name{concatenate_cols}
\alias{concatenate_cols}
\title{A function to concatenate the rows of supplied data frame columns}
\usage{
concatenate_cols(df, vars)
}
\arguments{
\item{df}{A tibble or data frame.}

\item{vars}{A vector of column names.}
}
\value{
A vector.
}
\description{
This function is similar to tidyr's unite() function and can be used to
concatenate the rows of supplied data frame columns.
}
\examples{
# Example with one key column (want to keep earliest date)
concat_example <- data.frame(id = c("id_3","id_1","id_2","id_2","id_1","id_3"),
date = as.Date(c("2024-05-01","2024-08-01","2024-02-14","2024-05-01","2024-05-01","2024-01-01")),
participantCode = c("NSF REU", "CAHSI", "DREU","DREU", "CAHSI", "NSF REU"))

concat_example$concat_cols <- concatenate_cols(df = concat_example,
vars = c("id", "date", "participantCode"))

}
\author{
Ama Nyame-Mensah
}
