% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_factor.R
\name{convert_to_factor}
\alias{convert_to_factor}
\title{Convert data.frame column to labelled factor.}
\usage{
convert_to_factor(df, var, codeBook, ...)
}
\arguments{
\item{df}{a data.frame or tibble.}

\item{var}{a character vector containing variable/column names to convert to a factor.}

\item{codeBook}{a data.frame or tibble containing: variable names, variable values, and value labels.}

\item{\dots}{additional parameters to pass to function.These include: variableName_col (an optional input
you can specify; variable names will be extracted from this column/field) and variableValues_col (an optional
input you can specify; variable values will be extracted from this column/field).
\itemize{
 \item \code{variableName_col}: An optional argument. This parameter is used to specify the name of the
 column/field in the codeBook to extract variable names from.
 \item \code{variableValues_col}: An optional argument. This parameter is used to specify the name of the
 column/field in the codeBook to extract variable values from.
 \item \code{valueLabels_col}: An optional argument. This parameter is used to specify the name of the
 column/field in the codeBook to extract value labels from.
}}
}
\value{
A data.frame or tibble.
}
\description{
This function leverages base's factor function to convert columns in a data.frame
or tibble to a labelled factor. NOTE: Defaults of addDelim are used for levels (,)
and labels (;).
}
\examples{
codeBook_race <- data.frame(variableName = "race_ethnicity", variableValues = "1,2,3,4",
valueLabels = "Black/African American;Asian/Asian American;Indigenous;Native Hawaiian/Pacific Islander")

race_tabl <- data.frame(race_ethnicity = sample(c(1:4, NA), size = 500, replace = TRUE))

convert_to_factor(df = race_tabl, var = "race_ethnicity", codeBook = codeBook_race)

}
\author{
Ama Nyame-Mensah
}
