% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spss_labelledVector_function.R
\name{create_spss_labelledVector_function}
\alias{create_spss_labelledVector_function}
\title{Function for quickly generating SPSS labelled vector using haven}
\usage{
create_spss_labelledVector_function(dataSet, variableName, ...)
}
\arguments{
\item{dataSet}{a tibble or data frame.}

\item{variableName}{name of a variable in a tibble or data frame to generate SPSS labelled vector for.}

\item{\dots}{additional parameters to pass to function. These include:
\itemize{
 \item \code{codeBook}: The name (string) of a tibble or data frame that contains codebook
 information.
 \item \code{codeBook_variableCol}: The name of the field/column in the codebook that contains
 variable names.
 \item \code{codeBook_questionLabelCol}: The name of the field/column in the codebook that contains
 question labels.
 \item \code{codeBook_valuesCol}: The name of the field/column in the codebook that contains variable
 values. Variable values should be in ONE cell and concatenated together using a comma (e.g., "1,2,3")).
 \item \code{codeBook_valueLabelsCol}: The name of the field/column in the codebook that contains value
 label. Value labels should be in ONE cell and concatenated together using a semi-colon (e.g., "One;Two;Three")).
 \item \code{variableValues}: A user-supplied vector of (variable) values. Variable values should be
 concatenated together using a comma (e.g., "1,2,3")).
 \item \code{valueLabels}: A user-supplied vector of value labels. Value labels should be concatenated
 together using a semi-colon (e.g., "One;Two;Three")).
 \item \code{questionLabel}: A user-supplied vector with a question label. Only one element should be
 supplied).
 \item \code{recodeMissing}: TRUE/FALSE. If set to TRUE, missing values are recoded. If set to FALSE
 values are not recoded.
 \item \code{recodeMissingValue} A user-supplied value to apply to missing values. If no missing value
 is supplied but recodeMissing is set to TRUE, then the default value "." is applied.
}}
}
\value{
A vector of labelled values.
}
\description{
This function can be used to quickly generate an SPSS labelled vector using
labelled_spss from the haven package. Users can either provide codebook inputs OR specify
inputs, but not both.
}
\examples{
# Example with supplied inputs, and missing is not recoded
spss_labelled_example <- data.frame(race_ethnicity = sample(c(1:4, NA),
size = 10, replace = TRUE))
create_spss_labelledVector_function(dataSet = spss_labelled_example,
variableName = "race_ethnicity",
variableValues = 1:4, valueLabels = c("Asian/Asian American", "Black/African American",
"Native Hawaiian/Pacific Islander", "Hispanic/Latinx"),
questionLabel = "What is your race/ethnicity?", recodeMissing = FALSE)

# Example with supplied codebook inputs
codebook_spss_labelled <- data.frame(variable_name = "sexNow",
question_label = "Do you currently describe yourself as male, female, or transgender?",
values = "1,2,3,4", value_labels = "Male;Female;Transgender;None of these")
spss_labelled_example2 <- data.frame(sexNow = sample(c(1:4, NA),
size = 10, replace = TRUE))
create_spss_labelledVector_function(dataSet = spss_labelled_example2,
codeBook = "codebook_spss_labelled", variableName = "sexNow",
codeBook_variableCol = "variable_name", codeBook_questionLabelCol = "question_label",
codeBook_valuesCol = "values", codeBook_valueLabelsCol = "value_labels",
recodeMissing = TRUE, recodeMissingValue = -9999)

}
\author{
Ama Nyame-Mensah
}
