% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distinct_rows_data}
\alias{distinct_rows_data}
\title{Distinct Rows Data Set}
\format{
A data frame with 6 rows and 2 variable(s):
\describe{
  \item{id}{participant identification numbers}
  \item{participantCode}{participant project codes}
  \item{recordedDate}{date survey response was recorded}
}
}
\usage{
distinct_rows_data
}
\description{
A data set containing generated id and participant code data.
}
\details{
Note: this data set is included for illustrative purposes only.

The variables are as follows:
}
\keyword{datasets}
