% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkMACS2}
\alias{.checkMACS2}
\title{Check that MACS2 can be found early on to save users time if MACS2 does not exist}
\usage{
.checkMACS2()
}
\description{
Check that MACS2 can be found early on to save users time if MACS2 does not exist
}
\keyword{internal}
