% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_punct.R
\name{escape_punct}
\alias{escape_punct}
\title{Function to escape punctuation for regex search .}
\usage{
escape_punct(x)
}
\arguments{
\item{x}{a character vector.}
}
\description{
Function to escape punctuation for regex search .
}
\examples{
stringSearch <- "All punctuations must be escaped. Will this function escape them? Yes."
escape_punct(stringSearch)

}
\author{
Ama Nyame-Mensah
}
