% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spss_label.R
\name{extract_spss_label}
\alias{extract_spss_label}
\title{Function for extracting variable labels from an SPSS data set}
\usage{
extract_spss_label(df, vars)
}
\arguments{
\item{df}{a tibble or data frame.}

\item{vars}{character with name of variable in df. Default is set to all, and all variable
labels will be extracted}
}
\value{
A data frame with two columns: variable names (column 1) and labels (column 2).
}
\description{
This function can be used to extract question/variable labels from
an SPSS data set.
}
\author{
Ama Nyame-Mensah
}
