% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBigWigs.R
\name{generateBigWigs}
\alias{generateBigWigs}
\title{Performs LSI, clustering and UMAP}
\usage{
generateBigWigs(
  archr.proj,
  groupBy,
  tileSize = 25,
  normMethod = "ReadsInTSS",
  maxCells = NULL
)
}
\arguments{
\item{archr.proj}{ArchRProject object}

\item{groupBy}{Metadata field to group cells by}

\item{tileSize}{Width of tiles that will section the genome}

\item{normMethod}{Normalization method for fragment pileups}

\item{maxCells}{Number of cells to use for each group.}
}
\value{
None
}
\description{
Sets global variables about specific genome build and number of threads to utilze on the machine.
}
\keyword{internal}
