% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gini_index.R
\name{get_gini_index_peak}
\alias{get_gini_index_peak}
\title{Get gini index for given peak}
\usage{
get_gini_index_peak(peak_index, peak_matrix)
}
\arguments{
\item{peak_matrix}{}
}
\description{
Get gini index for given peak
}
\keyword{internal}
