% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in_operator.R
\name{\%notin\%}
\alias{\%notin\%}
\title{This script creates the not in operator}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector of values.}

\item{y}{vector of values.}
}
\description{
This operator is the negated version of the %in% operator.
}
\examples{
values_vec <- 1:4
compare_vec <- sample(1:10, size = 2, replace = FALSE)

}
