% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCoverage.R
\name{groupCoverage}
\alias{groupCoverage}
\title{Creates coverages based on a variable of interest}
\usage{
groupCoverage(archr.proj, groupBy, ...)
}
\arguments{
\item{archr.proj}{ArchR project}

\item{groupBy}{Metadata field to group cells by}
}
\value{
archr.proj with coverages calculated for variable of interest
}
\description{
Creates coverages based on a variable of interest
}
\keyword{internal}
