% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlParser.R
\name{htmlParser}
\alias{htmlParser}
\title{Function to remove HTML entities from string}
\usage{
htmlParser(x)
}
\arguments{
\item{x}{a vector to remove HTML entities from}
}
\description{
This function removes HTML entities from string then standardizes spaces.
}
\author{
Ama Nyame-Mensah
}
