% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_to_gene.R
\name{make_p2g_table}
\alias{make_p2g_table}
\title{Construct peak to gene link table}
\usage{
make_p2g_table(archr.proj)
}
\arguments{
\item{marker.file}{}
}
\description{
Construct peak to gene link table
}
\keyword{internal}
