% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_return.R
\name{match_return}
\alias{match_return}
\title{Match and return value.}
\usage{
match_return(searchVal, lookUpDataFrame, lookUpCol, returnCol)
}
\arguments{
\item{searchVal}{A character vector containing a string or value to search for.}

\item{lookUpDataFrame}{A tibble or data frame.}

\item{lookUpCol}{The name of the column to search for searchVal.}

\item{returnCol}{The name of the column with the value to return.}
}
\value{
A vector.
}
\description{
This function is similar to xlookup in Excel.
It searches for a certain value in a column and then returns a value from a different column
in the same row. Note: NA is returned when no match is found
}
\examples{
race_key <- data.frame(value = 1:5, label = c("Black/African American", "Asian/Asian American",
"Indigenous", "Native Hawaiian/Pacific Islander", "No entry"))
race_tabl <- data.frame(race_ethnicity = sample(1:5, size = 10, replace = TRUE))
race_tabl$race_eth_label <- match_return(race_tabl$race_ethnicity, lookUpDataFrame = race_key,
lookUpCol = "value", returnCol = "label")

}
\author{
Ama Nyame-Mensah
}
