% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchingParams.R
\name{matchingParams}
\alias{matchingParams}
\title{Global matching parameters function}
\usage{
matchingParams(variableName, ...)
}
\arguments{
\item{variableName}{A variable name.}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{groupFlag.}: An optional parameter. that is a character vector containing the group
 or time flag pattern to identify each variable stem. The default is "_w". This default can be
 overridden by specifying another group or time flag pattern.
 }}
}
\description{
This function is primarily used to search for variables when generating data for analysis.
}
\author{
Ama Nyame-Mensah
}
