% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dataframes_lists.R
\name{merge_dataframes_lists}
\alias{merge_dataframes_lists}
\title{Function to merge list of data frames together on ONE key column.}
\usage{
merge_dataframes_lists(dfLists, ...)
}
\arguments{
\item{dfLists}{A list of data frames or tibbles}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{key (or id)}: An optional parameter. A user-specified key column to merge on.
 By default the column is set to "CERPID".
 }}
}
\description{
Function to merge list of data frames together on ONE key column.
}
\examples{
# Example
all_dfs <- list(df_1 = data.frame(id_key = c(3,1,2),
employee = c("heather", "burcin", "evelyn")),
df_2 = data.frame(title = c("director", "associate director",
"data manager"), id_key = c(1,3,2)))
merge_dataframes_lists(dfLists = all_dfs, keyColumn = "id_key")

}
\author{
Ama Nyame-Mensah
}
