% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNamesTabl.R
\name{mergeNamesTabl}
\alias{mergeNamesTabl}
\title{Function for strategically merging an auxiliary table with a key table table.}
\usage{
mergeNamesTabl(fullTabl, auxTabl, vars_to_add, ...)
}
\arguments{
\item{fullTabl}{A tibble or data frame; key table.}

\item{auxTabl}{A tibble or data frame; auxiliary table.}

\item{vars_to_add}{a character vector containing the names of the variables to look up.}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{set_commonkey}: An optional parameter you can specify. In the event the key variable in one
 table has a different name from the key variable in the second table, set this input to TRUE and then use
 the common_key input to set a new common key variable name.
 \item \code{fullTabl_variableCol}: An optional parameter (character vector) you can specify. In the event
 the key variable in the key table has a different name from the key variable in the auxiliary table,
 you must supply the field/variable name of the key variable to merge on.
 \item \code{auxTabl_variableCol}: An optional parameter (character vector) you can specify. In the event
 the key variable in the auxiliary table has a different name from the key variable in the key table,
 you must supply the field/variable name of the key variable to merge on.
  \item \code{common_key}: An optional parameter (character vector) you can specify. This will set
  a new common key variable name.
 }}
}
\value{
A data frame.
}
\description{
This function strategically merges two tables: 1) a key table and 2) an auxiliary table.
NOTE: The default key variable to merge on is "Variable". You must supply a character vector
of variables to lookup that exist in both the key and auxiliary tables. Key table data are
prioritized. In the event that a field/column is NA the column will be dropped. This function
was originally developed to be able to add data to an existing table that has NA's in certain
columns. If you accidentally add the same field/column with the same name and DIFFERENT values
to either table, a new row/observation will be added to the key table.
}
\examples{
full_tabl <- data.frame(orig = "aux_disability_grouped", timeFlag = "globalVar", Values = NA,
Labels = NA, string = NA, selectType = NA, orig_varLabel = NA)
aux_tabl <- data.frame(var_name = "aux_disability_grouped", Values = "0,1",
questionStem = "Do you have any disabilities?",
Labels = "No disabilities;One or more disabilities", selectType = "single choice",
groupType = "REU")
mergeNamesTabl(fullTabl = full_tabl, auxTabl = aux_tabl, vars_to_add = "aux_disability_grouped",
fullTabl_variableCol = "orig", auxTabl_variableCol = "var_name", set_commonkey = TRUE,
common_key = "varNAME")

}
\author{
Ama Nyame-Mensah
}
