% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_to_gene.R
\name{peak_to_gene}
\alias{peak_to_gene}
\title{Gets peak to gene links and merge with existing marker peaks}
\usage{
peak_to_gene(
  archr.proj = NULL,
  max_distance = 250000,
  reduced_dims = "LSI_Combined",
  useMatrix = "GeneExpressionMatrix",
  marker.file = file.path(getOutputDirectory(archr.proj), "MarkerPeaks", groupBy,
    paste0(groupBy, "_annotated_markerPeaks.tsv"))
)
}
\arguments{
\item{archr.proj}{}

\item{max_distance}{distance from the gene to search for linked peaks - default 250kb}

\item{reduced_dims}{reduced dimension parameter for p2gene links ArchR function, default "LSI_Combined"}

\item{useMatrix'}{matrix to use for constructing p2gene links, can be GeneExpressionMatrix (default) or GeneIntegrationMatrix}

\item{marker.file'}{annotated marker peaks file path}
}
\description{
Gets peak to gene links and merge with existing marker peaks
}
\keyword{internal}
