% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCaller.R
\name{peakAnnotation}
\alias{peakAnnotation}
\title{Annotate peaks using standard CERP format}
\usage{
peakAnnotation(archr.proj, groupBy, dataset, publish, ucsc.user, ucsc.session)
}
\arguments{
\item{archr.proj}{An already setup ArchR project or a name (string) to build a new ArchR project. If passing an already defined archr.proj then arrow.file.dir and cell.annotation.file will not be used.}

\item{groupBy}{Metadata field to group cells by}

\item{dataset}{Information about the dataset that these marker peaks belong to. E.g. Brain region, etc.}

\item{publish}{Set to 'TRUE' if you are ready to make the marker peak table viewable on Shiny.}

\item{ucsc.user}{A ucsc genome browser user/account which to contains the session to build links from.}

\item{ucsc.session}{A registred session for the UCSC genome browser user/account specified in \code{ucsc.user}.}
}
\value{
ArchRProject
}
\description{
Annotate peaks using standard CERP format
}
