% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCaller.R
\name{peakCaller}
\alias{peakCaller}
\title{Calls peaks currently utilizing ArchR}
\usage{
peakCaller(
  archr.proj,
  archr.genome,
  groupBy,
  dataset,
  archr.threads = 4,
  arrow.file.dir = NULL,
  cell.annotation.file = NULL,
  archr.clustering = FALSE,
  varFeatures = 15000,
  resolution = c(0.2, 1, 2),
  tileSize = 25,
  normMethod = "ReadsInTSS",
  maxCells = NULL,
  archr.visualize = FALSE,
  output.folder = NULL,
  publish = NULL,
  ucsc.user = NULL,
  ucsc.session = NULL
)
}
\arguments{
\item{archr.proj}{An already setup ArchR project or a name (string) to build a new ArchR project. If passing an already defined archr.proj then arrow.file.dir and cell.annotation.file will not be used.}

\item{archr.genome}{Valid genome name for ArchR: mm10, hg38}

\item{groupBy}{Metadata field to group cells by}

\item{dataset}{Information about the dataset that these marker peaks belong to. E.g. Brain region, etc.}

\item{archr.threads}{Number of threads to utilize, warning increases memory usage as well.}

\item{arrow.file.dir}{Directory where all the Arrow files are stored}

\item{cell.annotation.file}{File that contains sample_id and additional cell annotations to append to the ArchRProject}

\item{archr.clustering}{Boolean (T/F) determining if clustering/LSI should be run.}

\item{varFeatures}{Number of variable features to use for cluster}

\item{resolution}{Vector of clustering resolutions for multi-round clusters}

\item{tileSize}{Width of tiles that will section the genome}

\item{normMethod}{Normalization method for fragment pileups}

\item{maxCells}{Number of cells to use for each group.}

\item{archr.visualize}{Should marker peak visualizations be produced}

\item{output.folder}{If supplied we will save the marker peak table and ArchRProject to this location.}

\item{publish}{Set to 'TRUE' if you are ready to make the marker peak table viewable on Shiny.}

\item{ucsc.user}{A ucsc genome browser user/account which to contains the session to build links from.}

\item{ucsc.session}{A registred session for the UCSC genome browser user/account specified in \code{ucsc.user}.}
}
\value{
ArchRProject, markerPeaks
}
\description{
Sets global variables about specific genome build and number of threads to utilze on the machine.
}
