% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCalling.R
\name{peakCalling}
\alias{peakCalling}
\title{Calls peaks downstream of GroupCoverage based on variable of interest}
\usage{
peakCalling(archr.proj, groupBy, ...)
}
\arguments{
\item{archr.proj}{ArchR project with group coverages}

\item{groupBy}{Cell population of interest which should be in cellColData}

\item{...}{Additional arguments to addReproduciblePeakSet}
}
\value{
archr.proj with peaks called
}
\description{
Calls peaks downstream of GroupCoverage based on variable of interest
}
\keyword{internal}
