% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCaller.R
\name{peakRankeR_annotation}
\alias{peakRankeR_annotation}
\title{Rank peaks using the PeakRankR approach}
\usage{
peakRankeR_annotation(archr.proj, groupBy, marker.peak.table, bw.table = NULL)
}
\arguments{
\item{archr.proj}{An ArchR project after running through CERP}

\item{groupBy}{Metadata field to group cells by}

\item{marker.peak.table}{The annotated marker peak table from \code{peakAnnotation()}}

\item{bw.table}{A table with bigwig file names and corresponding annotations that match cell.population in marker.peak.table.}
}
\value{
ArchRProject
}
\description{
Adds new ranking columns to the peak annotation table
}
