% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCaller.R
\name{peakSpecificity}
\alias{peakSpecificity}
\title{Calls peaks currently utilizing ArchR}
\usage{
peakSpecificity(archr.proj, groupBy, max_sample_size = 500, numWorkers = 16)
}
\arguments{
\item{archr.proj}{An already setup ArchR project.}

\item{groupBy}{Metadata field to group cells by}

\item{max_sample_size}{}
}
\value{
ArchRProject
}
\description{
Sets global variables about specific genome build and number of threads to utilze on the machine.
}
