% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCaller.R
\name{peakToGene}
\alias{peakToGene}
\title{Calls peaks currently utilizing ArchR}
\usage{
peakToGene(
  archr.proj,
  max_distance,
  reduced_dims = "LSI_Combined",
  useMatrix = "GeneExpressionMatrix"
)
}
\arguments{
\item{archr.proj}{An already setup ArchR project.}

\item{max_distance}{ADD INFO}

\item{reduced_dims}{}

\item{useMatrix}{}
}
\value{
ArchRProject
}
\description{
Sets global variables about specific genome build and number of threads to utilze on the machine.
}
