% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicklyRenameVars.R
\name{quicklyRenameVars}
\alias{quicklyRenameVars}
\title{This function can be used to quickly rename variables}
\usage{
quicklyRenameVars(dataSet, codeBook, currentVar, oldNameCol, newNameCol)
}
\arguments{
\item{dataSet}{A tibble or data frame.}

\item{codeBook}{A tibble or data frame with at least two columns: one for old names and one for new names.}

\item{currentVar}{A character vector containing the names of the variables to update.}

\item{oldNameCol}{The name of the column in the code book containing old names to be replaced.}

\item{newNameCol}{The name of the column in the code book containing new names.}
}
\value{
A (character) vector of updated names.
}
\description{
This function allows the user to rename variables in a dataset. It requires a dataset and codebook
with two columns 1) one with old names, and 2) one with new names. NOTE: if the current variable name
is not found in the values supplied to "oldNameCol", the variable name will not be updated.
}
\examples{
quicklyRename_dat <- as.data.frame(matrix(data = NA, nrow = 1, ncol = 5))
quicklyRename_codebook <- data.frame(oldNames = paste0("V", 1:4),
newNames = paste0("question_",1:4))
names(quicklyRename_dat) <- quicklyRenameVars(dataSet = quicklyRename_dat,
codeBook = quicklyRename_codebook, currentVar = names(quicklyRename_dat),
oldNameCol = "oldNames", newNameCol = "newNames")

}
\author{
Ama Nyame-Mensah
}
