% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recode_apply_data}
\alias{recode_apply_data}
\title{Mapping Labels to Values Data Set}
\format{
A data frame with 6 rows and 2 variable(s):
\describe{
  \item{rank}{participant rank (number)}
}
}
\usage{
recode_apply_data
}
\description{
A data set containing generated rank values from 1 to 10.
}
\details{
Note: this data set is included for illustrative purposes only.

The variables are as follows:
}
\keyword{datasets}
