% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{rename_cols}
\alias{rename_cols}
\title{Function to strip group/time flags}
\usage{
rename_cols(x, ...)
}
\arguments{
\item{x}{A vector to split.}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
\item \code{groupFlag}: An optional parameter that is a character vector containing the group
or time flag pattern to identify each variable stem. The default is "_w". This default can be
overridden by specifying another group or time flag pattern. See example 2 below.
}}
}
\value{
A vector of values.
}
\description{
This is a generalized function for stripping group/time flags from variable names.
}
\examples{
# Example with default group/time flag
temp_dat <- data.frame(variable_name = c("sex_w1","sex_w2","sex"))
temp_dat$varStem <- rename_cols(x = temp_dat$variable_name)
temp_dat

# Example with user-specified group/time flag
names(yearBorn_data)
rename_cols(x = names(yearBorn_data), groupFlag = "_t")

}
\author{
Ama Nyame-Mensah
}
