% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gini_index.R
\name{sample_cells}
\alias{sample_cells}
\title{Subsamples cells for gini index calculation}
\usage{
sample_cells(group, cell_count_data = NULL, data = NULL)
}
\arguments{
\item{data}{}
}
\description{
Subsamples cells for gini index calculation
}
\keyword{internal}
