% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup ArchR for genome and threads}
\usage{
setup(archr.genome, archr.threads)
}
\arguments{
\item{archr.genome}{Valid genome name for ArchR: mm10, hg38}

\item{archr.threads}{Number of threads to utilize, warning increases memory usage as well.}
}
\value{
None
}
\description{
Sets global variables about specific genome build and number of threads to utilze on the machine.
}
\keyword{internal}
