% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexBirth_data}
\alias{sexBirth_data}
\title{Sex at Birth Multi-wave Data Set}
\format{
A data frame with 3 rows and 2 variables:
\describe{
 \item{id}{Participant id.}
 \item{sex_w1}{Time/Wave 1 variable indicating a students' self-identified sex at birth.}
 \item{sex_w2}{Time/Wave 2 variable indicating a students' self-identified sex at birth.}
}
}
\usage{
sexBirth_data
}
\description{
A data set containing information on students' self-identified sex at birth.
}
\details{
Note: this data set is included for illustrative purposes only.
}
\keyword{datasets}
