% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitChar.R
\name{splitChar}
\alias{splitChar}
\title{Split a sting on character (delimeter)}
\usage{
splitChar(vector, split, elementToKeep, ...)
}
\arguments{
\item{vector}{A vector to split.}

\item{split}{A delimiter to split string (e.g., comma, semi-colon).}

\item{elementToKeep}{the number of elements to retain.}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{addDelim}: An optional parameter you can specify when retaining more than one element.
 The default value is "".
 }}
}
\value{
A vector of values.
}
\description{
This is a generalized version of strsplit. You can retain 1 to n elements.
}
\examples{
# Example retaining the first element
city_data <- data.frame(city_state = c("Baltimore, Maryland", "Raleigh, North Carolina",
"Philadelphia, Pennsylvania"))
city_data$City <- splitChar(vector = city_data$city_state, split = ",", elementToKeep = 1)

# Example retaining the first and third elements (i.e., first and last names)
name_data <- data.frame(full_name = c("Linda E. Evans", "Akua Taylor Offor", "Heather M. Lutz"))
name_data$first_middle <- splitChar(vector = name_data$full_name,
split = " ", elementToKeep = c(1,3), addDelim = " ")

}
\author{
Ama Nyame-Mensah
}
