% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_match_extract.R
\name{str_match_extract}
\alias{str_match_extract}
\title{A function for extracting matched string from a vector.}
\usage{
str_match_extract(object, objectDictionary, ...)
}
\arguments{
\item{object}{a vector to search for match string}

\item{objectDictionary}{character vector to match in object}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{objectEscape}: An optional parameter that when set to TRUE escapes all punctuation in the object
 to search. Default is FALSE.
 \item \code{objectDictEscape}: An optional parameter that when set to TRUE escapes all punctuation in the
 object dictionary to search. Default is FALSE.
 \item \code{specialRun}: An optional argument that when set to TRUE will return all matched string
 separated by a semi-colon (e.g., match1; match2; match3). The default is FALSE, as only one match
 is typically expected.
 \item \code{invert}: An optional argument that when set to TRUE will return values for elements that
 do not match
 }}
}
\description{
This function is similar to str_extract from stringr. NOTE: case is ignored when matches are sought.
}
\examples{
# An example using text
string_to_match <- c("Choose up to 2 responses.", "Select all that apply.")
string_to_search <- data.frame(orig_varLabel = c("After you complete your current program,
what are your plans after graduation? Choose up to 2 responses.",
"How did you learn about the REU? Select all that apply.",
"This should return NA"))
string_to_search$extracted_string <- str_match_extract(object = string_to_search$orig_varLabel,
objectDictionary = string_to_match)

# An example returning multiple matched string
this_match <- c("Choose up to 2 responses.", "Select all that apply.", "Select all that apply")
string_to_search2 <- data.frame(orig_varLabel = c("After you complete your current program,
what are your plans after graduation? Choose up to 2 responses. Select all that apply",
"How did you learn about the REU? Select all that apply.", "This should return NA"))
string_to_search2$extracted_string <- str_match_extract(object = string_to_search2$orig_varLabel,
objectDictionary = this_match, specialRun = TRUE)

}
\author{
Ama Nyame-Mensah
}
