% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFlag_cols.R
\name{timeFlag_cols}
\alias{timeFlag_cols}
\title{Function to create indicator wave column}
\usage{
timeFlag_cols(x, ...)
}
\arguments{
\item{x}{a vector to split}

\item{\dots}{additional parameters to pass to function. These include:
\itemize{
 \item \code{groupFlag}: An optional parameter that is a character vector containing the group
 or time flag pattern to identify each variable stem. The default is "_w". This default can be
 overridden by specifying another group or time flag pattern. See example 2 below.
 }}
}
\value{
A vector of values. If a wave/grop flag is found, the flag is returned. Otherwise,
globalVar is returned.
}
\description{
This is a generalized function for extracting time or group flags for variables.
}
\examples{
# Example with default group/time flag
temp_dat <- data.frame(variable_name = c("sex_w1","sex_w2","sex"))
temp_dat$timeFlag <- timeFlag_cols(x = temp_dat$variable_name)
temp_dat

# Example with user-specified group/time flag
temp_dat2 <- data.frame(variable_name = c("highestDeg_it_group1", "csBackground_newProgLang",
"mentorSupport_group2"))
temp_dat2$groupFlag <- timeFlag_cols(x = temp_dat2$variable_name, groupFlag = "_group")
temp_dat2

}
\author{
Ama Nyame-Mensah
}
