% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlulist.R
\name{vlulist}
\alias{vlulist}
\title{Extract vector of labels}
\usage{
vlulist(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{addDelim}: An optional parameter you can specify when retaining more than one element.
 The default value is ";".
 }}
}
\value{
A character vector.
}
\description{
Extract vector of labels
}
\examples{
vlulist("one;two;three")

}
\author{
Ama Nyame-Mensah
}
