% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulist.R
\name{vulist}
\alias{vulist}
\title{Extract vector of values}
\usage{
vulist(x, ...)
}
\arguments{
\item{x}{A character vector}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{addDelim}: An optional parameter you can specify when retaining more than one element.
 The default value is ",".
 }}
}
\value{
A character vector.
}
\description{
Extract vector of values
}
\examples{
vulist("1,2,3")

}
\author{
Ama Nyame-Mensah
}
