% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_to_long_means.R
\name{wide_to_long_means}
\alias{wide_to_long_means}
\title{Calculate means for longitudinal variables with the same variable stem.}
\usage{
wide_to_long_means(dataSet, x, waves, ...)
}
\arguments{
\item{dataSet}{a tibble or data frame.}

\item{x}{a character vector of variable stems.}

\item{waves}{the number of waves/measurement occasions. Note: Wave numbers must be whole numbers and
correspond to the wave/group flag that appears in the data set (e.g., variable_w1 corresponds to wave
1, whereas variable_w3, corresponds to wave 3).}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{groupFlag}: An optional parameter that is a character vector containing the group
 or time flag pattern to identify each variable stem. The default is "_w". This default can be
 overridden by specifying another group or time flag pattern. See example 2 below.
 \item \code{listWise}: An optional parameter that when set to TRUE, listwise deletion is used.
 \item \code{wave_col_names}: Optional parameters that can be used to pass new field/column names to
 the table. For example, if you have two waves of data and you do not want pre_n and post_n to be
 printed, imply add wave_col_name1 = "wave1" and add wave_col_name2 = "wave2" and wave1_n and wave_2
 will be printed, respectively. See example 2 below for an example of how to use this parameter.
 }}
}
\value{
A tibble.
}
\description{
This function is a wrapper for calculating wide to long means for many variables with the same (variable) stem.
NOTE: By default, listwise deletion is set to TRUE (removing any records with NA values) and the default
wave/group flag is _w. Both of these can be modified using the listWise and groupFlag inputs, respectively
(see below).
}
\examples{
# Example with default group/time flag and listWise (deletion)
wide_tabl <- data.frame(belong_outsider_w1 = sample(1:5, replace = TRUE, size = 10),
belong_outsider_w2 = sample(c(NA,1:5), replace = TRUE, size = 10))
wide_to_long_means(x = "belong_outsider", dataSet = wide_tabl, waves = c(1,2))

# Example with user-specified group/time flag, listWise (deletion) set to FALSE,
# custom table headers are used, and multiple variables are considered
wide_n_tabl2 <- data.frame(belong_outsider_time1 = sample(1:5, replace = TRUE, size = 10),
belong_outsider_time3 = sample(c(NA,1:5), replace = TRUE, size = 10),
belong_welcomed_time1 = sample(c(NA,1:5), replace = TRUE, size = 10),
belong_welcomed_time3 = sample(c(NA,1:5), replace = TRUE, size = 10))
wide_to_long_means(x = c("belong_welcomed","belong_outsider"), dataSet = wide_n_tabl2,
waves = c(1L,3L), groupFlag = "_time",listWise = FALSE,wave_col_names = c("Wave1",
"Wave3"))

}
\author{
Ama Nyame-Mensah
}
