% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_to_long_n.R
\name{wide_to_long_n}
\alias{wide_to_long_n}
\title{Calculate n sizes for longitudinal variables with the same (variable) stem.}
\usage{
wide_to_long_n(x, dataSet, waves, ...)
}
\arguments{
\item{x}{a character vector of variable stems.}

\item{dataSet}{a tibble or data frame.}

\item{waves}{the number of waves/measurement occasions.}

\item{\dots}{Additional parameters to pass to function. These include:
\itemize{
 \item \code{groupFlag}: An optional parameter that is a character vector containing the group
 or time flag pattern to identify each variable stem. The default is "_w". This default can be
 overridden by specifying another group or time flag pattern. See example 2 below.
 \item \code{listWise}: An optional parameter that when set to TRUE, listwise deletion is used.
 \item \code{wave_col_names}: Optional parameters that can be used to pass new field/column names to
 the table. For example, if you have two waves of data and you do not want pre_n and post_n to be
 printed, imply add wave_col_name1 = "wave1" and add wave_col_name2 = "wave2" and wave1_n and wave_2
 will be printed, respectively. See example 2 below for an example of how to use this parameter.
 }}
}
\value{
A tibble.
}
\description{
This function is a wrapper for calculating wide to long n sizes of many variables with the same (variable)
stem. NOTE: By default, listwise deletion is set to TRUE (removing any records with NA values) and the default
wave/group flag is _w. Both of these can be modified using the listWise and groupFlag inputs, respectively
(see below).
}
\examples{
# Example with default group/time flag and listWise (deletion)
wide_n_tabl <- data.frame(belong_outsider_w1 = sample(1:5, replace = TRUE, size = 10),
belong_outsider_w2 = sample(c(NA,1:5), replace = TRUE, size = 10))
wide_to_long_n(x = "belong_outsider", dataSet = wide_n_tabl, waves = c(1,2))

# Example with user-specified group/time flag, listWise (deletion) set to FALSE,
# custom table headers are used, and multiple variables are considered
wide_n_tabl2 <- data.frame(belong_outsider_time1 = sample(1:5, replace = TRUE, size = 10),
belong_outsider_time3 = sample(c(NA,1:5), replace = TRUE, size = 10),
belong_welcomed_time1 = sample(c(NA,1:5), replace = TRUE, size = 10),
belong_welcomed_time3 = sample(c(NA,1:5), replace = TRUE, size = 10))
wide_to_long_n(x = c("belong_welcomed","belong_outsider"), dataSet = wide_n_tabl2,
waves = c(1,3), groupFlag = "_time",listWise = FALSE,wave_col_name1 = "Wave1",
wave_col_name2 = "Wave3")

}
\author{
Ama Nyame-Mensah
}
