% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yearBorn_data}
\alias{yearBorn_data}
\title{Year Born Multi-wave Data Set}
\format{
A data frame with 3 rows and 2 variables:
\describe{
 \item{id}{Participant id.}
 \item{yearBorn_t1}{Time/Wave 1 variable indicating when a student was born.}
 \item{yearBorn_t3}{Time/Wave 3 variable indicating when a student was born.}
}
}
\usage{
yearBorn_data
}
\description{
A data set containing information on the year students were born.
}
\details{
Note: this data set is included for illustrative purposes only.
}
\keyword{datasets}
