% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spss_label.R
\name{extract_spss_label}
\alias{extract_spss_label}
\title{Function for extracting question/variable labels for variables/columns from an SPSS data set}
\usage{
extract_spss_label(dataSet, variableName)
}
\arguments{
\item{dataSet}{a tibble or data frame.}

\item{variableName}{character with name of variable in dataSet.}
}
\value{
A data frame with two columns: variable names (column 1) and labels (column 2).
}
\description{
Function for extracting question/variable labels for variables/columns from an SPSS data set
}
\author{
Ama Nyame-Mensah
}
