% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in_operator.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negation of in}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{input vector to compare}

\item{table}{values to compare against}
}
\description{
This operator is the negated version of the %in% operator.
}
\examples{

a <- c(1,2,3)
30 \%notin\% a

}
