% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerPeaks.R
\name{markerPeaks}
\alias{markerPeaks}
\title{Function to calculate and export marker peaks. Exports}
\usage{
markerPeaks(archr.proj, groupBy, archr.visualize = TRUE)
}
\arguments{
\item{archr.proj}{ArchR object with peaks called}

\item{groupBy}{Metadata field to group cells by}

\item{archr.visualize}{Should marker peak visualizations be produced}
}
\value{
archr.proj, statistically significant peaks, peak matrix
}
\description{
Function to calculate and export marker peaks. Exports
}
\keyword{internal}
