% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerPeaks.R
\name{markerPeaks}
\alias{markerPeaks}
\title{Function to calculate and export marker peaks. Exports}
\usage{
markerPeaks(archr.proj, groupBy)
}
\arguments{
\item{archr.proj}{ArchR object with peaks called}

\item{groupBy}{Metadata field to group cells by}
}
\value{
marker peak table
}
\description{
Function to calculate and export marker peaks. Exports
}
\keyword{internal}
