% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCaller.R
\name{peakCaller}
\alias{peakCaller}
\title{Calls peaks currently utilizing ArchR}
\usage{
peakCaller(
  archr.proj,
  archr.genome,
  groupBy,
  archr.threads = 4,
  arrow.file.dir = NULL,
  cell.annotation.file = NULL,
  archr.clustering = FALSE,
  varFeatures = 15000,
  resolution = c(0.2, 1, 2),
  tileSize = 25,
  normMethod = "ReadsInTSS",
  maxCells = NULL,
  archr.visualize = TRUE,
  output.folder = NULL
)
}
\arguments{
\item{archr.proj}{An already setup ArchR project or a name (string) to build a new ArchR project. If passing an already defined archr.proj then arrow.file.dir and cell.annotation.file will not be used.}

\item{archr.genome}{Valid genome name for ArchR: mm10, hg38}

\item{groupBy}{Metadata field to group cells by}

\item{archr.threads}{Number of threads to utilize, warning increases memory usage as well.}

\item{arrow.file.dir}{Directory where all the Arrow files are stored}

\item{cell.annotation.file}{File that contains sample_id and additional cell annotations to append to the ArchRProject}

\item{archr.clustering}{Boolean (T/F) determining if clustering/LSI should be run.}

\item{varFeatures}{Number of variable features to use for cluster}

\item{resolution}{Vector of clustering resolutions for multi-round clusters}

\item{tileSize}{Width of tiles that will section the genome}

\item{normMethod}{Normalization method for fragment pileups}

\item{maxCells}{Number of cells to use for each group.}

\item{archr.visualize}{Should marker peak visualizations be produced}

\item{output.folder}{If supplied we will save the marker peak table and ArchRProject to this location.}
}
\value{
ArchRProject, markerPeaks
}
\description{
Sets global variables about specific genome build and number of threads to utilze on the machine.
}
